updates as `_data` becomes pd.DataFrame not list. created standalone `datagrid.py` monkeypatch file.
